added scroll down behavior
