+++ src/_scss/pages/about/aboutPage.scss
@@ -39,6 +39,7 @@
 
         ul {
             padding: 0;
+            scroll-behavior: smooth;
         }
 
         li {
