+++ src/js/components/about/navTest/About.jsx
@@ -3,10 +3,16 @@
  * Created by Mike Bray 11/20/2017
  **/
 
-import React from 'react';
+import React, {useEffect, useState} from 'react';
 
 import PageWrapper from 'components/sharedComponents/PageWrapper';
 import { aboutPageMetaTags } from 'helpers/metaTagHelper';
+import { stickyHeaderHeight } from 'dataMapping/stickyHeader/stickyHeader';
+import { getStickyBreakPointForSidebar } from 'helpers/stickyHeaderHelper';
+import { scrollToY } from 'helpers/scrollToHelper';
+import { useQueryParams } from 'helpers/queryParams';
+import { throttle } from "lodash";
+import { useHistory } from "react-router-dom";
 
 import AboutContent from './AboutContent';
 import InPageNav from "../../inPageNav/InPageNav";
@@ -89,17 +95,52 @@ const aboutSections = [
 ];
 
 
-const About = () => (
-    <PageWrapper
-        pageName="About"
-        classNames="usa-da-about-page"
-        metaTagProps={aboutPageMetaTags}
-        title="About">
-            <InPageNav sections={aboutSections} />
+const About = () => {
+    const query = useQueryParams();
+    const [activeSection, setActiveSection] = useState(query.section || 'mission');
+    const history = useHistory();
+
+    const jumpToSection = (section = '') => {
+        // if (!find(sections, { section })) { // not a known page section
+        //     return;
+        // }
+        const sectionDom = document.querySelector(`#about-${section}`);
+        if (!sectionDom) return;
+        const conditionalOffset = window.scrollY < getStickyBreakPointForSidebar() ? stickyHeaderHeight : 10;
+        const sectionTop = (sectionDom.offsetTop - stickyHeaderHeight - conditionalOffset);
+        scrollToY(sectionTop + 15, 700);
+        setActiveSection(section);
+    };
+
+
+    useEffect(throttle(() => {
+        // prevents a console error about react unmounted component leak
+        let isMounted = true;
+        if (isMounted) {
+            const urlSection = query.section;
+            if (urlSection) {
+                jumpToSection(urlSection);
+                // remove the query param from the url after scrolling to the given section
+                history.replace(`/about`);
+            }
+        }
+        return () => {
+            isMounted = false;
+        };
+    }, 100), [history, query.section]);
+
+    return (
+        <PageWrapper
+            pageName="About"
+            classNames="usa-da-about-page"
+            metaTagProps={aboutPageMetaTags}
+            title="About">
+            <InPageNav sections={aboutSections} jumpToSection={jumpToSection} />
             <main id="main-content" className="main-content">
                 <AboutContent />
             </main>
-    </PageWrapper>
-);
+        </PageWrapper>
+    );
+};
 
 export default About;
